third_party/blink/web_tests/platform/generic/fast/xpath/nsresolver-bad-object-expected.txt

CONSOLE ERROR: XPathNSResolver does not have a lookupNamespaceURI method.
SUCCESS
